server merge

+++ daily_health_server/src/main/java/com/example/daily_health_server/model/User.java
@@ -8,15 +8,15 @@ import javax.persistence.Id;
 import lombok.Getter;
 import lombok.Setter;
 
-@Getter
-@Setter
+@Getter @Setter
 @Entity
 public class User {
 	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long userid;
+	private int userid;
 	private String username;
 	private String email;
-	private Long height;
-	private Long weight;
+	private int height;
+	private int weight;
 	private String gender;
+
 }
